Time: 0 ms (100.00%) | Memory: 8.8 MB (92.95%) - LeetSync

+++ 58-length-of-last-word/length-of-last-word.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int lengthOfLastWord(string s) {
-        int i = static_cast<int>(s.size()) - 1;
+        short i = static_cast<short>(s.size()) - 1;
 
         while (i >= 0 && s[i] == ' ') {
             --i;
